convert login and otp page to client components
